Add box collider to cactus parallax
Added cactus box collider and object-dependent position getter function. The position for spawning will depend on the type of the obstacle, whether it is a cactus or an eagle or the reward. For cactus, the y position will be fixed to the ground. For all other objects, the y position will be randomized but with a cutoff, below which it should not spawn such object