Move albatross parallax objects into the environment folder
Right now, all the prefabs have been disabled, except for the eagle. When palying, the cactus, the albatross, and the eagle show up. But if the eagle prefab is disabled, it will not show up on the screen even though it is being spawned. The prefab object does not get re-enabled automatically as it should. Need to look into this issue further and force enable the prefabs when spawning programatically. Also the eagle and the albatross can spwan on top of one another. Need to make sure this never happens